requirements.txt
# Core dependencies
google-generativeai==0.8.3
fastapi==0.104.1
uvicorn==0.24.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Data processing
python-dateutil==2.8.2
pydantic==2.5.0

# Environment management
python-dotenv==1.0.0
